add all tests

+++ app/src/main/java/com/example/spa_wb_junior_devmeetingapp/ui/screens/registration/verificationScreen/VerificationViewModel.kt
@@ -56,7 +56,7 @@ class VerificationViewModel(
     fun onDoneKeyboardPressed(navigate: () -> Unit) {
         val pinCode = _uiState.value.pinCode
         viewModelScope.launch {
-            val isPinCodeValid = pinCodeVerificationUseCase.execute(pinCode.toInt())
+            val isPinCodeValid = pinCodeVerificationUseCase.execute(pinCode)
             when(isPinCodeValid){
                 true -> {
                     navigate()
